simulator/requirements.txt
paho-mqtt>=2.1.0
kafka-python>=2.0.2
